% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{.plot.CMAs}
\alias{.plot.CMAs}
\title{<defs>\n', # don't display it yet...}
\usage{
.plot.CMAs(cma, patients.to.plot = NULL, duration = NA,
  align.all.patients = FALSE, align.first.event.at.zero = TRUE,
  show.period = c("dates", "days")[2], period.in.days = 90,
  show.legend = TRUE, legend.x = "right", legend.y = "bottom",
  legend.bkg.opacity = 0.5, legend.cex = 0.75, legend.cex.title = 1,
  cex = 1, cex.axis = 0.75, cex.lab = 1, cex.title = 1.5,
  show.cma = TRUE, xlab = c(dates = "Date", days = "Days"),
  ylab = c(withoutCMA = "patient", withCMA = "patient (& CMA)"),
  title = c(aligned = "Event patterns (all patients aligned)", notaligned
  = "Event patterns"), col.cats = rainbow,
  unspecified.category.label = "drug", medication.groups = NULL,
  lty.event = "solid", lwd.event = 2, pch.start.event = 15,
  pch.end.event = 16, show.event.intervals = TRUE,
  plot.events.vertically.displaced = TRUE, print.dose = FALSE,
  cex.dose = 0.75, print.dose.col = "black",
  print.dose.outline.col = "white", print.dose.centered = FALSE,
  plot.dose = FALSE, lwd.event.max.dose = 8,
  plot.dose.lwd.across.medication.classes = FALSE,
  col.na = "lightgray", col.continuation = "black",
  lty.continuation = "dotted", lwd.continuation = 1,
  print.CMA = TRUE, CMA.cex = 0.5, plot.CMA = TRUE,
  plot.CMA.as.histogram = TRUE, plot.partial.CMAs.as = c("stacked",
  "overlapping", "timeseries")[1],
  plot.partial.CMAs.as.stacked.col.bars = "gray90",
  plot.partial.CMAs.as.stacked.col.border = "gray30",
  plot.partial.CMAs.as.stacked.col.text = "black",
  plot.partial.CMAs.as.timeseries.vspace = 7,
  plot.partial.CMAs.as.timeseries.start.from.zero = TRUE,
  plot.partial.CMAs.as.timeseries.col.dot = "darkblue",
  plot.partial.CMAs.as.timeseries.col.interval = "gray70",
  plot.partial.CMAs.as.timeseries.col.text = "firebrick",
  plot.partial.CMAs.as.timeseries.interval.type = c("none", "segments",
  "arrows", "lines", "rectangles")[2],
  plot.partial.CMAs.as.timeseries.lwd.interval = 1,
  plot.partial.CMAs.as.timeseries.alpha.interval = 0.25,
  plot.partial.CMAs.as.timeseries.show.0perc = TRUE,
  plot.partial.CMAs.as.timeseries.show.100perc = FALSE,
  plot.partial.CMAs.as.overlapping.alternate = TRUE,
  plot.partial.CMAs.as.overlapping.col.interval = "gray70",
  plot.partial.CMAs.as.overlapping.col.text = "firebrick",
  CMA.plot.ratio = 0.1, CMA.plot.col = "lightgreen",
  CMA.plot.border = "darkgreen", CMA.plot.bkg = "aquamarine",
  CMA.plot.text = CMA.plot.border, highlight.followup.window = TRUE,
  followup.window.col = "green", highlight.observation.window = TRUE,
  observation.window.col = "yellow", observation.window.density = 35,
  observation.window.angle = -30, observation.window.opacity = 0.3,
  show.real.obs.window.start = TRUE, real.obs.window.density = 35,
  real.obs.window.angle = 30, alternating.bands.cols = c("white",
  "gray95"), rotate.text = -60, bw.plot = FALSE,
  min.plot.size.in.characters.horiz = 0,
  min.plot.size.in.characters.vert = 0, max.patients.to.plot = 100,
  suppress.warnings = FALSE, export.formats = NULL,
  export.formats.fileprefix = "AdhereR-plot",
  export.formats.height = NA, export.formats.width = NA,
  export.formats.save.svg.placeholder = TRUE,
  export.formats.directory = NA, generate.R.plot = TRUE, ...)
}
\description{
<defs>\n', # don't display it yet...
}
