% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matching_function.R
\docType{data}
\name{durcomp.hospitalisation}
\alias{durcomp.hospitalisation}
\title{Example hospitalization events for 10 patients.}
\format{A data frame with 28 rows and 3 variables:
\describe{
  \item{ID}{\emph{integer} here; patient unique identifier. Can also
  be \emph{string}}.
  \item{DATE.IN}{\emph{Date} here;the start of the hospitalization period, by default in the
  yyyy-mm-dd format.Can also be \emph{string}.}
  \item{DATE.OUT}{\emph{Date};the end of the hospitalization period, by default in the
  yyyy-mm-dd format. Can also be \emph{string}.}
}}
\usage{
durcomp.hospitalisation
}
\description{
A sample dataset containing hospitalization periods (one per row) for 10 patients
over a period of roughly 18 months (28 events in total).
This is the appropriate format to compute event durations with the
\code{compute_event_durations} function. Each row represent an individual hospitalization period
of a patient for whom event durations should be calculated. Besides hospitalizations, this could
cover other situations where patients may not use their own supply, e.g. during incarcerations.
All column names must match the format provided in this example.
}
\keyword{datasets}
