% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{get.event.plotting.area}
\alias{get.event.plotting.area}
\title{Get the actual plotting area.}
\usage{
get.event.plotting.area(plot.type = c("baseR", "SVG")[1])
}
\value{
A numeric vector with components \emph{x.min}, \emph{x.max},
\emph{y.min} and \emph{y.max}, or \code{NULL} in case of error.
}
\description{
Returns the actual plotting area rectangle.
}
\details{
This is intended for advanced users only.
}
