% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adherer.R
\name{getEventsToEpisodesMapping}
\alias{getEventsToEpisodesMapping}
\title{getEventsToEpisodesMapping}
\usage{
getEventsToEpisodesMapping(x)
}
\arguments{
\item{x}{Either a \code{data.frame} as returned by \code{compute.treatment.episodes()}
or an \code{CMA_per_episode object}.}
}
\value{
The mapping between events and episodes, if it exists either as the
attribute \code{mapping.episodes.to.events} of the \code{data.frame} or as the
\code{mapping.episodes.to.events} component of the \code{CMA_per_episode object}
object, or \code{NULL} otherwise.
}
\description{
This function returns the event-to-episode mapping, if this information exists.
}
\details{
There are cases where it is interesting to know which events belong to which
episodes and which events have been actually used in computing the simple CMA
for each episode.
This information can be returned by \code{compute.treatment.episodes()} and
\code{CMA_per_episode()} if the parameter \code{return.mapping.events.episodes}
is set to \code{TRUE}.
}
