% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adherer.R
\name{getMGs}
\alias{getMGs}
\title{Access the medication groups of a CMA object.}
\usage{
getMGs(x)
}
\arguments{
\item{x}{a CMA object.}
}
\value{
a \emph{list} with two members:
\itemize{
 \item \code{defs} A \code{data.frame} containing the names and definitions of
 the medication classes; please note that there is an extra class
 \emph{__ALL_OTHERS__} containing all the observations not selected by any of
 the explicitly given medication classes.
 \item \code{obs} A \code{logical matrix} where the columns are the medication
 classes (the last being \emph{__ALL_OTHERS__}), and the rows the observations in
 the x's data; element \eqn{(i,j)} is \code{TRUE} iff observation \eqn{j} was
 selected by medication class \eqn{i}.
}
}
\description{
Retrieve the medication groups and the observations they refer to (if any).
}
