% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{last.plot.get.info}
\alias{last.plot.get.info}
\title{Access last adherence plot info.}
\usage{
last.plot.get.info()
}
\value{
A \code{list} (possibly empty) contaning one named element for each type
of plot produced (currently only \emph{baseR} and \emph{SVG}). Each may contain
shared and specific fields concerning:
\itemize{
 \item the values of the parameters with which \code{plot()} was invoked.
 \item actual plot size and other characteristics.
 \item actual title, axis names and labels and their position and size.
 \item legend size, position and size and position of its components.
 \item expanded \code{cma$data} contaning, for each event, info about its
 plotting, including the corresponding fullow-uo and observation windows,
 event start and end, dose text (if any) and other graphical elements.
 \item position, size of the partial CMAs (if any) and of their components.
 \item position, size of the plotted CMAs (if any) and of their components.
 \item rescaling function(s) useful for mapping events to plotting coordinates.
 \item for \emph{baseR} only, the original and used par() environment.
}
}
\description{
Returns the full info the last adherence plot, to be used to modify and/or to
add new elements to this plot.
}
\details{
This is intended for advanced users only.
It may return \code{NULL} if no plotting was generated yet, but if one was, a
list contaning one named element for each type of plot produced (currently only
\emph{baseR} and \emph{SVG} are used).
Besides the shared elements (see the returned value), there are specific ones
as well.
For \emph{baseR}, the members \emph{old.par} and \emph{used.par} contain the
original (pre-plot) \code{par()} environment and the one used within
\code{plot()}, respectively. Before \code{plot()} does anything, it saves the
environment into \emph{old.par}, then it adjusts it in various ways and it saves
it to \emph{used.par}, does its job, then, just beofre returning, restores back
\emph{old.par}. So, beofre adding/chaning anything to the plot, please make sure
to reinstate \emph{used.par}, do the plotting, and restore back \emph{old.par}
(see examples).
}
\examples{
cma7 <- CMA7(data=med.events[med.events$PATIENT_ID \%in\% c(1,2),],
             ID.colname="PATIENT_ID",
             event.date.colname="DATE",
             event.duration.colname="DURATION",
             event.daily.dose.colname="PERDAY",
             medication.class.colname="CATEGORY",
             followup.window.start=0,
             followup.window.start.unit="days",
             followup.window.duration=2*365,
             followup.window.duration.unit="days",
             observation.window.start=30,
             observation.window.start.unit="days",
             observation.window.duration=365,
             observation.window.duration.unit="days",
             date.format="\%m/\%d/\%Y",
             summary="Base CMA");
plot(cma7);
tmp <- last.plot.get.info();
names(tmp);
tmp$baseR$legend$box; # legend position and size
head(tmp$baseR$cma$data); # events + plotting info
par(tmp$baseR$used.par); # restore the plotting environment
rect(tmp$baseR$.map.event.x(270), tmp$baseR$.map.event.y(1-0.5),
     tmp$baseR$.map.event.x(900), tmp$baseR$.map.event.y(nrow(tmp$baseR$cma$data)+0.5),
     col=adjustcolor("blue",alpha.f=0.5), border="blue");
par(tmp$baseR$old.par); # restore the pre-plotting environment
}
