% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adherer.R
\docType{data}
\name{med.events.ATC}
\alias{med.events.ATC}
\title{Example of medication events with ATC codes.}
\format{
A data frame with 1564 rows and 7 variables:
\describe{
  \item{PATIENT_ID}{the patient unique identifier.}
  \item{DATE}{the medication event date.}
  \item{DURATION}{the duration in days.}
  \item{PERDAY}{the daily dosage.}
  \item{CATEGORY}{the ATC code.}
  \item{CATEGORY_L1}{explicitation of the first field of the ATC code (e.g.,
  "A"="ALIMENTARY TRACT AND METABOLISM").}
  \item{CATEGORY_L2}{explicitation of the first and second fields of the ATC
  code (e.g., "A02"="DRUGS FOR ACID RELATED DISORDERS").}
}
}
\usage{
med.events.ATC
}
\description{
An artificial dataset containing medication events (one per row) for 16
patients (1564 events in total), containing ATC codes. This dataset is
derived from the \code{durcomp} datasets using the \code{compute_event_durations}
function. See @med.events for more details.
}
\keyword{datasets}
