% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adherer.R
\name{plot_interactive_cma}
\alias{plot_interactive_cma}
\title{Interactive exploration and CMA computation.}
\usage{
plot_interactive_cma(...)
}
\arguments{
\item{...}{Parameters to be passed to \code{plot_interactive_cma()} in package
\code{AdhereRViz}.}
}
\value{
Nothing
}
\description{
Interactively plot a given patient's data, allowing the real-time exploration
of the various CMAs and their parameters.
It can use \code{Rstudio}'s \code{manipulate} library or \code{Shiny}.
}
\details{
This is merely a stub for the actual implementation in package
\code{AdhereRViz}: it just checks if this package is installed and functional,
in which case it calls the actual implementation, otherwise warns the user that
\code{AdhereRViz} must be instaled.
}
\examples{
\dontrun{
plot_interactive_cma(med.events,
                     ID.colname="PATIENT_ID",
                     event.date.colname="DATE",
                     event.duration.colname="DURATION",
                     event.daily.dose.colname="PERDAY",
                     medication.class.colname="CATEGORY");}
}
\seealso{
Function \code{\link[AdhereR]{plot_interactive_cma}} in package
\code{AdhereRViz}.
}
