% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matching_function.R
\docType{data}
\name{durcomp.hospitalisation}
\alias{durcomp.hospitalisation}
\title{Example special periods for 10 patients.}
\format{A data frame with 28 rows and 3 variables:
\describe{
  \item{ID}{\emph{Integer} here; patient unique identifier. Can also
  be \emph{string}.}
  \item{DATE.IN}{\emph{Date} here;the start of the hospitalization period, by default in the
  yyyy-mm-dd format.Can also be \emph{string}.}
  \item{DATE.OUT}{\emph{Date};the end of the hospitalization period, by default in the
  yyyy-mm-dd format. Can also be \emph{string}.}
}}
\usage{
durcomp.hospitalisation
}
\description{
A sample dataset containing special periods (one per row) for 10 patients
over a period of roughly 18 months (28 events in total).
This is the appropriate format to compute event durations with the
\code{compute_event_durations} function. Each row represents an individual special period of type
"hospitalization" of a patient for whom event durations should be calculated.
Besides hospitalizations, this could cover other situations
where medication use may differ, e.g. during incarcerations or holidays.
All column names must match the format provided in this example.
}
\keyword{datasets}
