% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call-adherer-external.R
\name{getCallerWrapperLocation}
\alias{getCallerWrapperLocation}
\title{getCallerWrapperLocation.}
\usage{
getCallerWrapperLocation(callig.platform = c("python3")[1],
  full.path = FALSE)
}
\arguments{
\item{callig.platform}{A \emph{string} specifying the desired wrapper.
Currently it can be "python3".}

\item{full.path}{A \emph{logical} specifying if the returned path should
also include the wrapper's main file name.}
}
\value{
The full path to the requested wrapper or NULL if none exists.
}
\description{
This function returns the full path to where the various \code{wrappers} that
can call \code{AdhereR} are installed.
}
\details{
In most cases, these wrappers are one or more files in the calling language
that may be directly used as such.
For more details see the vignette describing the included reference
\code{Python 3} wrapper.
}
